requirements.txt
jira
slackclient
requests
urllib
